ioready-epoll: assert return of epoll_ctl
These asserts guard against programming errors of the library user.

A bug where an fd was reused and closed again in user code
was basically found because some epoll_ctl with EPOLL_CTL_MOD
failed. The tscb library assumed that the kernel had still some
epoll information for a fd. But the application closed that fd without
informing the tscb library about it.

So if one the assertions happens, check if the correct
order of disconnect first, close second is used.

In the above error case the close was called from some component which
closed the same fd twice with a long interval, and the fd was reused
at the second call of close.

To disable the asserts call configure with --disable-assert